+++ src/Definition/PaginatorAwareInterface.php
@@ -2,17 +2,18 @@
 
 namespace Knp\Bundle\PaginatorBundle\Definition;
 
-use Knp\Component\Pager\Paginator;
+use Knp\Component\Pager\PaginatorInterface;
 
 /**
  * Interface PaginatorAwareInterface.
  *
- * PaginatorAwareInterface should be implemented by classes that depend on a KnpPaginator service.
+ * PaginatorAwareInterface can be implemented by classes that depend on a KnpPaginator service.
+ * You should avoid this solution: use autowiring instead.
  */
 interface PaginatorAwareInterface
 {
     /**
      * Sets the KnpPaginator instance.
      */
-    public function setPaginator(Paginator $paginator): self;
+    public function setPaginator(PaginatorInterface $paginator): self;
 }